Django framwork/django-begin/requirements.txt
django
django-crispy-forms
python-slugify
djangorestframework